+++ plugins/module_utils/device_inventory/delete.py
@@ -16,18 +16,19 @@ import requests
 # fmt: on
 
 
+# TODO: replace extra filter....
 def find_device_for_deletion(module_params: dict, http_session: requests.session, endpoint: str):
     """Find the object we intend to delete"""
-    if module_params["device_type"].upper() == "FTD":
+    if module_params.get("device_type").upper() == "FTD":
         extra_filter = "AND (deviceType:FTDC)"
     else:
-        extra_filter = f"AND (deviceType:{module_params['device_type'].upper()})"
-    module_params["filter"] = module_params["device_name"]
+        extra_filter = f"AND (deviceType:{module_params.get('device_type').upper()})"
+    module_params["filter"] = module_params.get(["device_name"])
     device_list = gather_inventory(module_params, http_session, endpoint, extra_filter=extra_filter)
     if len(device_list) < 1:
-        raise DeviceNotFound(f"Cannot delete {module_params['device_name']} - device by that name not found")
+        raise DeviceNotFound(f"Cannot delete {module_params.get('device_name')} - device by that name not found")
     elif len(device_list) > 1:
-        raise TooManyMatches(f"Cannot delete {module_params['device_name']} - more than 1 device matches name")
+        raise TooManyMatches(f"Cannot delete {module_params.get('device_name')} - more than 1 device matches name")
     else:
         return device_list[0]
 
@@ -37,13 +38,13 @@ def delete_device(module_params: dict, http_session: requests.session, endpoint:
     try:
         device = find_device_for_deletion(module_params, http_session, endpoint)
         working_set(http_session, endpoint, device["uid"])
-        if module_params["device_type"].upper() == "ASA" or module_params["device_type"].upper() == "IOS":
+        if module_params.get("device_type").upper() == "ASA" or module_params.get("device_type").upper() == "IOS":
             response = CDORequests.delete(
                 http_session, f"https://{endpoint}", path=f"{CDOAPI.DEVICES.value}/{device['uid']}"
             )
             return response
 
-        elif module_params["device_type"].upper() == "FTD":
+        elif module_params.get("device_type").upper() == "FTD":
             cdfmc = get_cdfmc(http_session, endpoint)
             cdfmc_specific_device = get_specific_device(http_session, endpoint, cdfmc["uid"])
             data = {