+++ CHANGELOG.rst
@@ -17,3 +17,10 @@ v1.0.2
 ======
 - Moved all inventory function into the inventory module and updated the argument spec
 - Updated the sample playbooks in /docs
+
+v1.1.0
+======
+- Breaking Change: Reworked onboarding and inventory
+- New module: Deploy - to view and deploy staged ASA changes
+- See new sample playbooks and sample inventory
+- Updated the sample playbooks in /docs